3/27까지 커밋

+++ src/main/java/com/board/spring03/repository/BoardRepository.java
@@ -1,12 +1,12 @@
 package com.board.spring03.repository;
 
-import com.board.spring03.momain.BoardWrite;
+import com.board.spring03.model.BoardWrite;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface BoardRepository extends JpaRepository<BoardWrite, Integer> {
+public interface BoardRepository extends JpaRepository<BoardWrite, Long> {
     List<BoardWrite> findAllByOrderByModifiedAtDesc();
 }
