+++ src/main/java/com/board/spring03/service/BoardService.java
@@ -1,7 +1,7 @@
 package com.board.spring03.service;
 
 import com.board.spring03.repository.BoardRepository;
-import com.board.spring03.momain.BoardWrite;
+import com.board.spring03.model.BoardWrite;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,13 +16,16 @@ public class BoardService {
 
 
     public List<BoardWrite> boardList (){
+
         return boardRepository.findAllByOrderByModifiedAtDesc();
     }
 
     @Transactional
-    public BoardWrite boardViewer(Integer id){
+    public BoardWrite boardViewer(Long id){
         return boardRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("존재하지 않는 게시글 입니다.")
         );
     }
+
+
 }
